pyproject.toml
[tool.poetry]
name = "weather checker application"
version = "0.1.0"
description = "An application for checking the weather worldwide"
authors = ["Alon Shmueli <alonshmu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
streamlit = "^1.35.0"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
